+++ src/main/java/com/example/flightmanager/adapter/SqlFlightRepository.java
@@ -1,21 +1,13 @@
-package com.example.flightmanager.repository;
+package com.example.flightmanager.adapter;
 
 import com.example.flightmanager.model.Flight;
+import com.example.flightmanager.repository.FlightRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
-public interface FlightRepository {
-
-    List<Flight> findAll();
-
-    Optional<Flight> findById(Integer id);
-
-    Flight save(Flight entity);
-
-    void delete(Flight entity);
-
+public interface SqlFlightRepository extends FlightRepository, JpaRepository<Flight, Integer> {
+    @Override
     List<Flight> findByRouteContainingAndDateAfterAndAvailableSeatsGreaterThanEqual(String route, LocalDateTime date, int availableSeats);
 }
